+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii 2 ActiveRecord Variation extension Change Log
 1.0.1 under development
 -----------------------
 
+- Bug #12: `VariationBehavior` does not use `ActiveRecord::getRelation()` while retrieving relation instance (klimov-paul)
 - Bug #11: Relation declared via `VariationBehavior::hasDefaultVariationRelation()` does not support `LEFT JOIN` (klimov-paul)